Fixed DynamicEntityService, moved all logic from DynamicEntityRepository to it. Some exceptional conditions are described. Added new songstants with error messages
