+++ Octopus.Core.Common/Constants/ErrorMessages.cs
@@ -14,6 +14,8 @@
 
         public const string DynamicServiceException = "An exception was thrown during the work with dynamic service exception. The inner exception is ";
 
+        public const string DatabaseProviderException = "A connection provider cannot be created for this database type - ";
+
         public const string UnhandledException = "An unhandled exception was thrown. The inner exception is ";
     }
 }
