+++ Octopus.Core.Common/Exceptions/DatabaseProviderException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Octopus.Core.Common.Exceptions
+{
+    public class DatabaseProviderException : Exception
+    {
+        public DatabaseProviderException(string msg) : base(msg)
+        {
+            
+        }
+    }
+}
